requirements-docs.txt
# Requirements for building hosted Sphinx builds
# NOTE these requirements are duplicates of requirements.txt
mock


Django==1.6.5       # rq.filter: >=1.6,<1.7

# DPaW custom
# https://static.dpaw.wa.gov.au/static/py/dpaw-utils/dist/dpaw-utils-0.3a11.tar.gz
# provides:
# django-confy>=1.0.4
django-extensions==1.6.5
# django-redis==3.6.2
# psycopg2==2.5.2

# Database, models
django-contrib-postgres
django-model-utils
django-polymorphic==0.5.3
djorm-pgarray==1.2
djorm-ext-pgarray
South==0.8.4
#-e git://github.com/lukesneeringer/django-pgfields.git#egg=django-pgfields

diff-match-patch==20110725.1

django-bootstrap-pagination==1.5.0
django-contrib-comments==1.5
django-cors-headers==0.12
django-extra-views==0.6.4

django-guardian==1.2.0
django-immutablemodel==0.3.4
django-mail-templated==2.6.2
django-reversion==1.8.0
django-taggit==0.12
factory-boy==2.3.1
fancy-tag==0.2.0
pystache==0.5.4
python-magic==0.4.6
semantic-version==2.3.0

# API
djangorestframework==2.3.13
django-rest-swagger==0.1.14
django-filter==0.11.0

# Export to CSV
django_tablib==3.2

# Comments
# django-disqus
# provides:
# six

# DevOps
argparse
django-confy
django-wsgiserver>=0.8.0rc1
Fabric
gunicorn
wsgiref==0.1.2
pipdeptree
django-appconf==0.6

# Docs, images, thumbnails
django-resized
Sphinx==1.5.5
pycallgraph
graphviz
mock
Pillow
pydot
pyparsing
django-resized

# Forms and widgets
django-crispy-forms==1.4.0
django-leaflet==0.13.2
Django-Select2==4.2.2
django-smart-selects
django-uni-form==0.9.0
humanize==0.5

# Latex, Markdown, JSON, XML, YAML
django-markup-deprecated==0.0.3
lxml==3.7.3
beautifulsoup4==4.5.3
django-tinymce==1.5.2
html2text==2016.9.19
html5lib==0.999999999
pyenchant==1.6.8
pypandoc==1.4
Unidecode==0.04.16

# Static files
#static==1.0.2
dj-static
django-compressor==1.6
django-htmlmin==0.7.0
unipath>=1.1

# Task queues
django-celery==3.1.10
django-redis-sessions==0.4.0

# Templating
django-grappelli>=2.6.5
django-bootstrap3>=5.3.1,<6.0
django-braces>=1.8.0
jinja2>=2.7.3
webtemplate-dpaw==0.3

# Transitions
django-fsm==2.3.0
django-fsm-log

# Testing and Debugging
django-debug-toolbar==1.1
django-dtpanel-htmltidy
django-nose==1.3
django-pdb
python-coveralls
#mixer
selenium
ipdbplugin

# Linters
flake8
flake8-docstrings
flake8_import_order
flake8-todo
pep8-naming
radon
